Moving Property and PropertiesSet out of SBOM
This will remove a risk of circular dependencies
when SPDXSbom is introduced.
